Fix Administrator checking and include elevation prompt.
IsUserAnAdmin() just checks whether the user is a member of the Administrator group. On modern Windows (since Vista), you can be a member of the Administrator group, but normal processes run with a reduced privilege level by default. So this is a poor determination of whether your Python process actually has full admin privileges. IsNTAdmin() is a much better determinant. If you right-click on something and do "Run As Administrator", IsNTAdmin() will return True; otherwise it returns False, even if you are in the Administrator group. This will resolve permissions problems when executing certain functionality.